@@ -14,20 +14,20 @@ import 'package:flutter_tts/flutter_tts.dart';
 enum TtsState { playing, stopped, paused, continued }
 
 class EmotionRecognizer extends StatefulWidget {
-  const EmotionRecognizer({key});
+  const EmotionRecognizer({super.key});
 
   @override
   State<EmotionRecognizer> createState() => _EmotionRecognizerState();
 }
 
 class _EmotionRecognizerState extends State<EmotionRecognizer> {
-  CameraController _controller;
-  FlutterTts flutterTts;
+  late CameraController _controller;
+  late FlutterTts flutterTts;
   double volume = 2.5;
   double pitch = 1.0;
   double rate = 0.5;
-  Future<void> _initializeControllerFuture;
-  CameraDescription camera;
+  late Future<void> _initializeControllerFuture;
+  late CameraDescription camera;
   TtsState ttsState = TtsState.stopped;
 
   get isPlaying => ttsState == TtsState.playing;
@@ -257,7 +257,8 @@ class _EmotionRecognizerState extends State<EmotionRecognizer> {
 class DisplayPictureScreen extends StatelessWidget {
   final String imagePath;
 
-  const DisplayPictureScreen({key, this.imagePath});
+  const DisplayPictureScreen({super.key, required this.imagePath});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
