@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_barcode_scanner/flutter_barcode_scanner.dart';
 
 class BarcodeReader extends StatelessWidget {
-  const BarcodeReader({key});
+  const BarcodeReader({super.key});
 
   @override
   Widget build(BuildContext context) {
